fixed get/setRowProperties; added filterEditor
