+++ src/behaviors/Column.js
@@ -315,8 +315,11 @@ Column.prototype = {
      */
     getCellEditorAt: function(cellEvent) {
         var columnIndex = this.index,
+
             rowIndex = cellEvent.gridCell.y,
+
             editorName = cellEvent.properties.editor,
+
             options = Object.create(cellEvent, {
                 format: {
                     // `options.format` is a copy of the cell's `format` property which is:
@@ -327,6 +330,7 @@ Column.prototype = {
                     value: cellEvent.properties.format
                 }
             }),
+
             cellEditor = this.dataModel.getCellEditorAt(columnIndex, rowIndex, editorName, options);
 
         if (cellEditor && !cellEditor.grid) {
@@ -399,7 +403,7 @@ function resolveCalculator(calculator) {
     return calculators[key];
 }
 
-Column.prototype.mixIn(require('./cellProperties').mixin);
+Column.prototype.mixIn(require('./cellProperties').columnMixin);
 Column.prototype.mixIn(require('./columnProperties').mixin);
 
 module.exports = Column;
